+++ src/app/dataset-flow/flows-table/flows-table.helpers.ts
@@ -10,6 +10,7 @@ import {
     FlowStatus,
     FlowSummaryDataFragment,
     FlowTimingRecords,
+    FlowTriggerBatchingRuleBuffering,
 } from "src/app/api/kamu.graphql.interface";
 import { MaybeNull } from "src/app/interface/app.types";
 import AppValues from "src/app/common/values/app.values";
@@ -368,7 +369,7 @@ export class FlowTableHelpers {
                 return "waiting for a throttling condition";
 
             case "FlowStartConditionReactive":
-                return "waiting for a reactive condition";
+                return `waiting for input data: ${startCondition.accumulatedRecordsCount}/${(startCondition.activeBatchingRule as FlowTriggerBatchingRuleBuffering).minRecordsToAwait}`;
 
             case "FlowStartConditionExecutor": {
                 return "waiting for a free executor";