+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] 
+### Added
+- Flows table: visualizing the latest batching condition statictics
 ### Changed
 - Metadata tab: combined "Dataset Type", "Name", and "Owner" into one property "Dataset"
-
 
 ## [0.56.0] - 2025-08-28
 ### Added